% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiPvr-package.R
\docType{package}
\name{EpiPvr-package}
\alias{EpiPvr}
\alias{EpiPvr-package}
\title{The EpiPvr package}
\description{
Tools for epidemic parameter estimation using Stan.
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan.
R package version 2.32.6. https://mc-stan.org
}
\seealso{
Useful links:
\itemize{
  \item \url{https://CRAN.R-project.org/package=EpiPvr}
}

}
\author{
\strong{Maintainer}: Ruairi Donnelly \email{rd501@cam.ac.uk} [copyright holder]

}
\keyword{internal}
