% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AP_insect_simulator.R
\name{gen_exp}
\alias{gen_exp}
\title{Safely generate random exponential data}
\usage{
gen_exp(n, rin, virusType)
}
\arguments{
\item{n}{Numeric: number of points to sample.}

\item{rin}{Numeric: event rate.}

\item{virusType}{Character: SPT or PT plant virus type.}
}
\value{
Numeric vector: n random exponential number.
}
\description{
A helper function only needed within AP_insect_simulator that generates random numbers from an exponential distribution
using the base/stats R function \code{stats::runif}.
This function simulates a set of random numbers with a fixed rate safely (ie for different input cases stemming from virus types).
}
\keyword{internal}
