% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inoc_durtn_calculator.R
\name{inoc_durtn_calculator}
\alias{inoc_durtn_calculator}
\title{Helper function for effective inoculation period calculator}
\usage{
inoc_durtn_calculator(lmark_in, sim_Ts_in, isitVerbose)
}
\arguments{
\item{lmark_in}{Numeric vector of length 3 specifying assay durations.}

\item{sim_Ts_in}{Numeric vector of length 4 specifying sampled event times (hours) related to virus parameters (rates hr^-1) (rates per hour)}

\item{isitVerbose}{Logical (default: \code{FALSE}). If \code{TRUE}, prints information from nested functions.}
}
\value{
Numeric: effective inoculation duration.
}
\description{
Helper function for effective inoculation period calculator
}
\keyword{internal}
